[refactor] 챌린지 통계 로직 서비스 단에서 분리
